style(clippy): crate-wide clippy cleanup - exit 0
Resolve the pre-existing deny-level correctness errors that left `cargo
clippy` red (exit 101): approx_constant in near_pole_tan test helpers
(-> std::f64::consts::LOG2_10, bit-identical for the test inputs) and an
intentional a-a identity check (allow(eq_op) + note). Apply all
machine-applicable suggestions across lib + tests (unnecessary_fallible_
conversions, unnecessary_cast, redundant parens, ...).

clippy now exits 0. Remaining advisory warnings (~160, mostly macro-generated
spans rustfix cannot rewrite + intentional should_implement_trait) tracked for
a later deliberate pass. Full suite + golden 264 green.

+++ tests/d38_max_scale_log_panic.rs
@@ -32,6 +32,6 @@ fn log10_strict_on_saturated_d38_scale38() {
 /// failure to the kernel rather than `from_f64`'s saturation policy.
 #[test]
 fn log10_strict_on_near_max_d38_scale38() {
-    let v = D38::<38>::from_bits(decimal_scaled::Int::<2>::try_from((i128::MAX) as i128).unwrap());
+    let v = D38::<38>::from_bits(decimal_scaled::Int::<2>::try_from(i128::MAX).unwrap());
     let _ = v.log10_strict();
 }
