+++ tests/ulp_proptest.rs
@@ -165,7 +165,7 @@ proptest! {
     #![proptest_config(proptest_config("exp_of_ln_roundtrip"))]
     #[test]
     fn exp_of_ln_roundtrip(raw in positive_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let y = x.ln_strict().exp_strict();
         let xb = i128::from(x.to_bits());
         let yb = i128::from(y.to_bits());
@@ -178,7 +178,7 @@ proptest! {
     #![proptest_config(proptest_config("ln_of_exp_roundtrip"))]
     #[test]
     fn ln_of_exp_roundtrip(raw in ln_exp_domain()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let y = x.exp_strict().ln_strict();
         let xb = i128::from(x.to_bits());
         let yb = i128::from(y.to_bits());
@@ -191,11 +191,11 @@ proptest! {
     #![proptest_config(proptest_config("sin2_plus_cos2_is_one"))]
     #[test]
     fn sin2_plus_cos2_is_one(raw in real_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let s = x.sin_strict();
         let c = x.cos_strict();
         let sum = s * s + c * c;
-        let one = D::from_bits(decimal_scaled::Int::<2>::try_from((ONE) as i128).unwrap());
+        let one = D::from_bits(decimal_scaled::Int::<2>::try_from(ONE).unwrap());
         let tol = round_trip_tol(ONE);
         prop_assert_within!("sin²+cos²",
                             i128::from(sum.to_bits()), i128::from(one.to_bits()),
@@ -207,7 +207,7 @@ proptest! {
     #![proptest_config(proptest_config("sqrt_squared_roundtrip"))]
     #[test]
     fn sqrt_squared_roundtrip(raw in nonneg_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let r = x.sqrt_strict();
         let back = r * r;
         let xb = i128::from(x.to_bits());
@@ -221,7 +221,7 @@ proptest! {
     #![proptest_config(proptest_config("cbrt_cubed_roundtrip"))]
     #[test]
     fn cbrt_cubed_roundtrip(raw in real_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let r = x.cbrt_strict();
         let back = r * r * r;
         let xb = i128::from(x.to_bits());
@@ -235,7 +235,7 @@ proptest! {
     #![proptest_config(proptest_config("atan_of_tan_roundtrip"))]
     #[test]
     fn atan_of_tan_roundtrip(raw in open_quarter_pi()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let y = x.tan_strict().atan_strict();
         let xb = i128::from(x.to_bits());
         let yb = i128::from(y.to_bits());
@@ -248,7 +248,7 @@ proptest! {
     #![proptest_config(proptest_config("tanh_atanh_roundtrip"))]
     #[test]
     fn tanh_atanh_roundtrip(raw in open_interval_neg1_to_1()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let y = x.atanh_strict().tanh_strict();
         let xb = i128::from(x.to_bits());
         let yb = i128::from(y.to_bits());
@@ -261,7 +261,7 @@ proptest! {
     #![proptest_config(proptest_config("sin_odd_symmetry"))]
     #[test]
     fn sin_odd_symmetry(raw in real_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let lhs = x.sin_strict();
         let rhs = -((-x).sin_strict());
         prop_assert_within!("sin(-x)=-sin(x)", i128::from(lhs.to_bits()), i128::from(rhs.to_bits()),
@@ -274,7 +274,7 @@ proptest! {
     #![proptest_config(proptest_config("cos_even_symmetry"))]
     #[test]
     fn cos_even_symmetry(raw in real_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let lhs = x.cos_strict();
         let rhs = (-x).cos_strict();
         prop_assert_within!("cos(-x)=cos(x)", i128::from(lhs.to_bits()), i128::from(rhs.to_bits()),
@@ -287,7 +287,7 @@ proptest! {
     #![proptest_config(proptest_config("atan_odd_symmetry"))]
     #[test]
     fn atan_odd_symmetry(raw in real_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let lhs = x.atan_strict();
         let rhs = -((-x).atan_strict());
         prop_assert_within!("atan(-x)=-atan(x)", i128::from(lhs.to_bits()), i128::from(rhs.to_bits()),
@@ -300,7 +300,7 @@ proptest! {
     #![proptest_config(proptest_config("cbrt_odd_symmetry"))]
     #[test]
     fn cbrt_odd_symmetry(raw in real_x()) {
-        let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+        let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
         let lhs = x.cbrt_strict();
         let rhs = -((-x).cbrt_strict());
         prop_assert_within!("cbrt(-x)=-cbrt(x)", i128::from(lhs.to_bits()), i128::from(rhs.to_bits()),
@@ -354,7 +354,7 @@ mod wide_witness {
         #![proptest_config(config("d76_sqrt_witness"))]
         #[test]
         fn d76_sqrt_agrees_with_d38(raw in nonneg_x()) {
-            let n = N::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let n = N::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let w: W = n.into();
             let nb = i128::from(n.sqrt_strict().to_bits());
             let wb = w.sqrt_strict().to_bits().to_i128_checked()
@@ -369,7 +369,7 @@ mod wide_witness {
         #![proptest_config(config("d76_exp_witness"))]
         #[test]
         fn d76_exp_agrees_with_d38(raw in moderate_real()) {
-            let n = N::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let n = N::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let w: W = n.into();
             let nb = i128::from(n.exp_strict().to_bits());
             let wb = w.exp_strict().to_bits().to_i128_checked()
@@ -438,7 +438,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_tie_sqrt_roundtrip"))]
         #[test]
         fn hard_tie_sqrt_roundtrip(raw in near_unit()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let r = x.sqrt_strict();
             let back = r * r;
             let xb = i128::from(x.to_bits());
@@ -472,7 +472,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_canc_exp_of_ln"))]
         #[test]
         fn hard_canc_exp_of_ln(raw in ln_just_above_one()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let y = x.ln_strict().exp_strict();
             let xb = i128::from(x.to_bits());
             let yb = i128::from(y.to_bits());
@@ -487,11 +487,11 @@ mod hard_inputs {
         #![proptest_config(config("hard_canc_pythag_tiny"))]
         #[test]
         fn hard_canc_pythag_tiny(raw in tiny_around_zero()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let s = x.sin_strict();
             let c = x.cos_strict();
             let sum = s * s + c * c;
-            let one = D::from_bits(decimal_scaled::Int::<2>::try_from((ONE) as i128).unwrap());
+            let one = D::from_bits(decimal_scaled::Int::<2>::try_from(ONE).unwrap());
             let tol = round_trip_tol(ONE);
             let diff = (i128::from(sum.to_bits()) - i128::from(one.to_bits())).abs();
             prop_assert!(diff <= tol,
@@ -528,7 +528,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_rred_sin_symmetry"))]
         #[test]
         fn hard_rred_sin_symmetry(raw in near_half_pi_multiples()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let lhs = x.sin_strict();
             let rhs = -((-x).sin_strict());
             let diff = (i128::from(lhs.to_bits()) - i128::from(rhs.to_bits())).abs();
@@ -542,11 +542,11 @@ mod hard_inputs {
         #![proptest_config(config("hard_rred_pythag"))]
         #[test]
         fn hard_rred_pythag(raw in near_quarter_pi_odd_multiples()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let s = x.sin_strict();
             let c = x.cos_strict();
             let sum = s * s + c * c;
-            let one = D::from_bits(decimal_scaled::Int::<2>::try_from((ONE) as i128).unwrap());
+            let one = D::from_bits(decimal_scaled::Int::<2>::try_from(ONE).unwrap());
             let tol = round_trip_tol(ONE);
             let diff = (i128::from(sum.to_bits()) - i128::from(one.to_bits())).abs();
             prop_assert!(diff <= tol,
@@ -575,7 +575,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_asym_exp_of_ln_small"))]
         #[test]
         fn hard_asym_exp_of_ln_small(raw in ln_near_zero()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let y = x.ln_strict().exp_strict();
             let xb = i128::from(x.to_bits());
             let yb = i128::from(y.to_bits());
@@ -590,7 +590,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_asym_atan_odd_huge"))]
         #[test]
         fn hard_asym_atan_odd_huge(raw in huge_real()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let lhs = x.atan_strict();
             let rhs = -((-x).atan_strict());
             let diff = (i128::from(lhs.to_bits()) - i128::from(rhs.to_bits())).abs();
@@ -633,7 +633,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_inv_atan_of_tan"))]
         #[test]
         fn hard_inv_atan_of_tan(raw in quarter_pi_inner()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let y = x.tan_strict().atan_strict();
             let xb = i128::from(x.to_bits());
             let yb = i128::from(y.to_bits());
@@ -648,7 +648,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_inv_sqrt_of_square"))]
         #[test]
         fn hard_inv_sqrt_of_square(raw in squares_of_inputs()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let r = x.sqrt_strict();
             let back = r * r;
             let xb = i128::from(x.to_bits());
@@ -686,7 +686,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_pp_sqrt_roundtrip"))]
         #[test]
         fn hard_pp_sqrt_roundtrip(raw in perfect_squares_jittered()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let r = x.sqrt_strict();
             let back = r * r;
             let xb = i128::from(x.to_bits());
@@ -702,7 +702,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_pp_cbrt_symmetry"))]
         #[test]
         fn hard_pp_cbrt_symmetry(raw in perfect_cubes_jittered()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let lhs = x.cbrt_strict();
             let rhs = -((-x).cbrt_strict());
             let diff = (i128::from(lhs.to_bits()) - i128::from(rhs.to_bits())).abs();
@@ -735,11 +735,11 @@ mod hard_inputs {
         #![proptest_config(config("hard_const_pythag"))]
         #[test]
         fn hard_const_pythag(raw in near_pi_constants()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let s = x.sin_strict();
             let c = x.cos_strict();
             let sum = s * s + c * c;
-            let one = D::from_bits(decimal_scaled::Int::<2>::try_from((ONE) as i128).unwrap());
+            let one = D::from_bits(decimal_scaled::Int::<2>::try_from(ONE).unwrap());
             let tol = round_trip_tol(ONE);
             let diff = (i128::from(sum.to_bits()) - i128::from(one.to_bits())).abs();
             prop_assert!(diff <= tol,
@@ -754,7 +754,7 @@ mod hard_inputs {
         fn hard_const_exp_of_ln(raw in near_e_unit()) {
             // Only positive inputs hit ln's domain.
             prop_assume!(raw > 0);
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let y = x.ln_strict().exp_strict();
             let xb = i128::from(x.to_bits());
             let yb = i128::from(y.to_bits());
@@ -791,7 +791,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_halv_atan_odd"))]
         #[test]
         fn hard_halv_atan_odd(raw in atan_halving_anchors()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let lhs = x.atan_strict();
             let rhs = -((-x).atan_strict());
             let diff = (i128::from(lhs.to_bits()) - i128::from(rhs.to_bits())).abs();
@@ -829,7 +829,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_exp_stage2_roundtrip"))]
         #[test]
         fn hard_exp_stage2_roundtrip(raw in near_k_ln2()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             // exp(x) might overflow at |x| > ~44 in storage; the
             // strategy keeps |k| <= 30 so we're well inside.
             let y = x.exp_strict().ln_strict();
@@ -862,7 +862,7 @@ mod hard_inputs {
         #![proptest_config(config("hard_tang_exp_of_ln"))]
         #[test]
         fn hard_tang_exp_of_ln(raw in tang_lookup_anchors()) {
-            let x = D::from_bits(decimal_scaled::Int::<2>::try_from((raw) as i128).unwrap());
+            let x = D::from_bits(decimal_scaled::Int::<2>::try_from(raw).unwrap());
             let y = x.ln_strict().exp_strict();
             let xb = i128::from(x.to_bits());
             let yb = i128::from(y.to_bits());
